+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add `Async::Signals.default` and `Async::Signals::Ignore` for selecting process signal handling based on the current thread.
+
 ## v0.1.0
 
   - Initial release.